requirements.txt
# Python dependencies
mcp[cli]
aioimaplib
beautifulsoup4
lxml
apscheduler
aiosqlite
python-dotenv
pydantic
